requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
beartype==0.14.1
beautifulsoup4==4.14.2
bitsandbytes==0.48.1
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
datasets==4.2.0
detoxify==0.5.2
dill==0.3.7
einops==0.7.0
evaluate==0.4.6
exceptiongroup==1.3.0
fancy-einsum==0.0.3
filelock==3.19.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2023.10.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.26.0
google-api-python-client==2.108.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.71.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.11
iniconfig==2.3.0
jaxtyping==0.3.3
Jinja2==3.1.6
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.15
networkx==3.3
nlp==0.4.0
numpy==1.26.4
openai>=1.0.0 # Added openai for evaluations
packaging==25.0
pandas==2.3.3
peft==0.17.1
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psutil==7.1.1
pyarrow==21.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.22
requests==2.32.5
rich==14.2.0
rsa==4.9.1
safetensors==0.6.2
seaborn==0.13.1
sentencepiece==0.2.1
sentry-sdk==2.42.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
sympy==1.13.1
tabulate==0.9.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
tqdm==4.67.1
transformer-lens==1.12.0
transformers==4.57.1
triton==3.1.0
trl==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wadler_lindig==0.1.7
wandb==0.22.2
xxhash==3.6.0
yarl==1.22.0
